# Remove redundant hashing from OrderedSet insertion