# Introduction
This pr implements a NoOp hasher for use within the `OrderedSet`'s element index. This is ordered set returns the last 8 (or less) bytes written to it. This should _only_ be used in the context of returning an already generated hash.


## Benchmark Results

This change has yielded a ~17% improvement to table_set generation.

before
```
Benchmarking lr1 table generation: Warming up for 3.0000 s
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 10.6s, or reduce sample count to 40.
lr1 table generation    time:   [105.49 ms 105.55 ms 105.63 ms]
                        change: [-2.4185% -2.3277% -2.2396%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 12 outliers among 100 measurements (12.00%)
  2 (2.00%) low severe
  1 (1.00%) low mild
  2 (2.00%) high mild
  7 (7.00%) high severe
```

after
```
Benchmarking lr1 table generation: Warming up for 3.0000 s
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 8.5s, or reduce sample count to 50.
lr1 table generation    time:   [86.610 ms 88.475 ms 90.487 ms]
Found 20 outliers among 100 measurements (20.00%)
  19 (19.00%) high mild
  1 (1.00%) high severe
```
# Linked Issues
#41 
# Dependencies

# Test
- [x] Tested Locally
- [x] Documented

# Review
- [x] Ready for review
- [x] Ready to merge

# Deployment